requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
matplotlib
python-dotenv
